Adds put and get tests for HDF5 files
This commit adds two files that demonstrate how to store HDF5 files in
DataSpaces and add a client API that allows to store and retrieve
arbitrary files to/from DataSpaces.

Here files are stored as one-dimension arrays in the spaces in the bound
[0, fsize] where fsize is the file size in bytes.
The client API is dcg_get_dims(const char* var_name, unsigned int ver,
int ndim, uint64_t *dims) queries the DHT to get the dimension of arbitrary
(single-object) variable in DataSpaces, which is required for the get
operation.